fix(evpn-bridge): move ipu to intel e2000
